Add class DatasetFileSink.
